requirements.txt
asgiref==3.3.4
boto3==1.17.78
botocore==1.20.78
cffi==1.14.5
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.1
django-environ==0.4.5
django-restframework==0.0.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.0
gunicorn==20.1.0
jmespath==0.10.0
jwt==1.2.0
Pillow==8.2.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
